1.12.2-PR support (and predicted 1.12.2 support)

+++ proxy/src/main/java/net/md_5/bungee/UserConnection.java
@@ -92,7 +92,7 @@ public final class UserConnection implements ProxiedPlayer
     /*========================================================================*/
     @Getter
     @Setter
-    private int sentPingId;
+    private Long sentPingId;
     @Getter
     @Setter
     private long sentPingTime;
